flask_app/requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
streamlit
tiktoken
openai
python-dotenv
azure-search-documents
azure-core
pyodbc
